{"schema":"decodex/commit/1","summary":"Split core support modules","authority":"manual"}
